layout/forms/test/test_bug378670.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=378670
-->
<head>
  <title>Test for Bug 378670</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=378670">Mozilla Bug 378670</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
Clicking on the select should not crash Mozilla
<select id="select">
<option>1</option>
<option>2</option>
</select>

<pre id="test">
<script>
document.getElementsByTagName('*')[3].addEventListener('popupshowing', function(e) {e.target.parentNode.removeChild(e.target) }, true);
</script>
<script type="application/javascript">

/** Test for Bug 378670 **/

SimpleTest.waitForExplicitFinish();

function clickit() {
  netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');
  var select = document.getElementById('select');
  var evt = document.createEvent("MouseEvents");
  evt.initMouseEvent("mousedown", true, true, window, 0, 0, 0, 0, 0, false, false, false, false, 0, null);
  select.dispatchEvent(evt);
  evt.initMouseEvent("mouseup", true, true, window, 0, 0, 0, 0, 0, false, false, false, false, 0, null);
  select.dispatchEvent(evt);
  evt.initMouseEvent("click", true, true, window, 0, 0, 0, 0, 0, false, false, false, false, 0, null);
  select.dispatchEvent(evt);

  setTimeout(finish, 200);

}

window.addEventListener('load', clickit, false);

function finish()
{
  ok(true, "This is a mochikit version of a crash test. To complete is to pass.");
  SimpleTest.finish();
}
</script>
</pre>
</body>
</html>
